feat: Indexing implementation (part)
Refs: #298

+++ evita_store/evita_store_server/src/main/java/io/evitadb/store/traffic/DiskRingBuffer.java
@@ -6,7 +6,7 @@
  *             |  __/\ V /| | || (_| | |_| | |_) |
  *              \___| \_/ |_|\__\__,_|____/|____/
  *
- *   Copyright (c) 2024
+ *   Copyright (c) 2024-2025
  *
  *   Licensed under the Business Source License, Version 1.1 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -307,7 +307,7 @@ public class DiskRingBuffer {
 		@Nonnull SessionLocation sessionLocation,
 		@Nonnull UUID sessionId,
 		@Nonnull OffsetDateTime created,
-		long durationInMillis,
+		int durationInMillis,
 		@Nonnull Set<TrafficRecordingType> recordingTypes,
 		int fetchCount,
 		int bytesFetchedTotal
